refactor: improve CLI output
