+++ crates/biome_cli/src/execute/migrate/eslint.rs
@@ -7,13 +7,13 @@ use biome_json_parser::JsonParserOptions;
 use biome_service::DynRef;
 use std::borrow::Cow;
 use std::path::{Path, PathBuf};
-use std::process::Command;
 use std::vec;
 
 use crate::diagnostics::MigrationDiagnostic;
 use crate::CliDiagnostic;
 
 use super::eslint_eslint;
+use super::node;
 
 /// This modules includes implementations for loading and deserializing an eslint configuration.
 ///
@@ -122,7 +122,7 @@ fn load_flat_config_data(
     path: &Path,
     console: &mut dyn Console,
 ) -> Result<eslint_eslint::FlatConfigData, CliDiagnostic> {
-    let NodeResolveResult { content, .. } = load_config_with_node(&path.to_string_lossy())?;
+    let node::Resolution { content, .. } = node::load_config(&path.to_string_lossy())?;
     let (deserialized, diagnostics) = deserialize_from_json_str::<eslint_eslint::FlatConfigData>(
         &content,
         JsonParserOptions::default(),
@@ -182,7 +182,7 @@ fn load_legacy_config_data(
             }
         },
         Some("js" | "cjs") => {
-            let NodeResolveResult { content, ..} = load_config_with_node(&path.to_string_lossy())?;
+            let node::Resolution { content, ..} = node::load_config(&path.to_string_lossy())?;
             deserialize_from_json_str::<eslint_eslint::LegacyConfigData>(
                 &content,
                 JsonParserOptions::default(),
@@ -214,66 +214,6 @@ fn load_legacy_config_data(
         Err(CliDiagnostic::MigrateError(MigrationDiagnostic {
             reason: "Could not deserialize the Eslint configuration file".to_string(),
         }))
-    }
-}
-
-#[derive(Debug)]
-struct NodeResolveResult {
-    /// Resolved path of the file
-    resolved_path: String,
-    /// File content
-    content: String,
-}
-
-/// Imports `specifier` using Node's `import()` or node's `require()` and
-/// returns the JSONified content of its default export.
-fn load_config_with_node(specifier: &str) -> Result<NodeResolveResult, CliDiagnostic> {
-    let content_output = Command::new("node")
-        .arg("--eval")
-        .arg(format!(
-            "import('{specifier}').then((c) => console.log(JSON.stringify(c.default)))"
-        ))
-        .output();
-    match content_output {
-        Err(_) => {
-            Err(CliDiagnostic::MigrateError(MigrationDiagnostic {
-                reason: "The `node` program doesn't exist or cannot be invoked by Biome.\n`node` is invoked to resolve ESLint configurations written in JavaScript.\nThis includes shared configurations and plugin configurations imported with ESLint's `extends`.".to_string()
-            }))
-        },
-        Ok(output) => {
-            let path_output = Command::new("node")
-                .arg("--print")
-                .arg(format!(
-                    "require.resolve('{specifier}')"
-                ))
-                .output();
-            let resolved_path = path_output.ok().map_or(String::new(), |path_output| String::from_utf8_lossy(&path_output.stdout).trim().to_string());
-            if !output.stderr.is_empty() {
-                // Try with `require` before giving up.
-                let output2 = Command::new("node")
-                    .arg("--print")
-                    .arg(format!(
-                        "JSON.stringify(require('{specifier}'))"
-                    ))
-                    .output();
-                if let Ok(output2) = output2 {
-                    if output2.stderr.is_empty() {
-                        return Ok(NodeResolveResult {
-                            content: String::from_utf8_lossy(&output2.stdout).to_string(),
-                            resolved_path,
-                        });
-                    }
-                }
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                return Err(CliDiagnostic::MigrateError(MigrationDiagnostic {
-                    reason: format!("`node` was invoked to resolve an ESLint configuration. This invocation failed with the following error:\n{stderr}")
-                }));
-            }
-            Ok(NodeResolveResult {
-                content: String::from_utf8_lossy(&output.stdout).to_string(),
-                resolved_path,
-            })
-        }
     }
 }
 
@@ -322,10 +262,10 @@ fn load_eslint_extends_config(
             }
         };
         // load ESLint preset
-        let Ok(NodeResolveResult {
+        let Ok(node::Resolution {
             content,
             resolved_path,
-        }) = load_config_with_node(&module_name)
+        }) = node::load_config(&module_name)
         else {
             return Err(CliDiagnostic::MigrateError(MigrationDiagnostic {
                 reason: format!(
@@ -358,10 +298,10 @@ fn load_eslint_extends_config(
         } else {
             EslintPackage::Config.resolve_name(name)
         };
-        let Ok(NodeResolveResult {
+        let Ok(node::Resolution {
             content,
             resolved_path,
-        }) = load_config_with_node(&module_name)
+        }) = node::load_config(&module_name)
         else {
             return Err(CliDiagnostic::MigrateError(MigrationDiagnostic {
                 reason: format!(
